requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
sentence-transformers
PyPDF2
faiss-cpu
